Выполнена работа над ошибками
