+++ ImageFeed/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -1,4 +1,9 @@
 {
+  "colors" : [
+    {
+      "idiom" : "universal"
+    }
+  ],
   "info" : {
     "author" : "xcode",
     "version" : 1
