+++ ImageFeed/Assets.xcassets/Images/favorites_active.imageset/Active@2x.png
