+++ ImageFeed/Assets.xcassets/Images/favorites_active.imageset/Active@3x.png
