+++ ImageFeed/Assets.xcassets/Images/favorites_active.imageset/Contents.json
@@ -1,17 +1,16 @@
 {
   "images" : [
     {
-      "filename" : "Vector.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "Vector (1).png",
+      "filename" : "Active@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "Vector (2).png",
+      "filename" : "Active@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
