+++ ImageFeed/Assets.xcassets/Images/favorites_no_active.imageset/Contents.json
@@ -1,17 +1,16 @@
 {
   "images" : [
     {
-      "filename" : "Active.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "Active@2x.png",
+      "filename" : "No Active@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "Active@3x.png",
+      "filename" : "No Active@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
