+++ ImageFeed/Assets.xcassets/Images/favorites_no_active.imageset/No Active@2x.png
