+++ ImageFeed/Assets.xcassets/Images/splash_screen.imageset/Vector.png
