+++ ImageFeed/Assets.xcassets/Images/splash_screen.imageset/Vector@2x.png
