+++ ImageFeed/Assets.xcassets/Images/splash_screen.imageset/Vector@3x.png
