--- ImageFeed/ImagesListCell.swift
@@ -1,9 +1,0 @@
-import UIKit
-
-final class ImagesListCell: UITableViewCell {
-    static let reuseIdentifier = "ImagesListCell"
-    
-    @IBOutlet var cellImage: UIImageView!
-    @IBOutlet var dataLabel: UILabel!
-    @IBOutlet var likeButton: UIButton!
-}
